Migrate to NixOS unstable
Switched the nixpkgs input to the `nixos-unstable` branch. This was
mainly done to get access to KWin 6.3, which includes a performance
improvement which might solve my lagspike stuttering problem when
playing games (I've tried it now and it might have improved things
slightly).

This required some changes due to removed/deprecated options:
- Removed the nix configuration from 'home.nix'. I don't think it
  actually did anything either way.
- Enabled `useFetchCargoVendor` when building the `rlr` package.
- Use the new dedicated `nerd-fonts.fira-code` package.
- Changed `hardware.pulseaudio.enable` -> `services.pulseaudio.enable`