Revert "Migrate to NixOS unstable"
This reverts commit 6fbd5f9405ea9825652ca14f49c35201fa9ea82d.

There seems to be a bug where the rendering of parts of the screen is
"corrupted". Sometimes it freezes the screen, sometimes also showing
more corrupted imagery.
One time it fully crashed KWin, after which the desktop environment
had restarted with a crash report window shown.
The problems seem to get worse when Zed is opened.

I haven't been able to figure out a way to solve it, so I'm just going
to go back to the stable version. I'll lose KDE Plasma 6.3.3, but that
beats my desktop environment crashing.

They weren't lying about it being unstable at least. :P