internal: Add .git-blame-ignore-revs
Added the previous commit which formatted the entire tree to the
revisions ignored by Git blame.

I left a comment with instructions for how to configure Git to use it.
To do so run:
```shell
git config --local blame.ignoreRevsFile .git-blame-ignore-revs
```

Also configured GitLens to respect it in the workspace's VSCode
settings.